requirements.txt
Django==5.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.4.0
reportlab==4.0.7
psycopg2-binary==2.9.10